+++ lib/flemme.dart
@@ -1,3 +1,4 @@
 library flemme;
 
+export 'src/config_file.dart';
 export 'src/widget_extension.dart';