+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.5
+* add withContainer(), withSizedBox(),withFractionnalySizedBox().
+* add default padding configuration, default backgroundColor for modal
+* add option on showHimSelf()
+* add option on listToRow & listToColumn
+
+
 ## 0.0.3
 * add documentation, example and complete read.me.